feat: move credits display to dev preferences
